Add files via upload
This Excel file consists of 3 sheets. The first sheet is the gridded multiplication table 1-12. The 2nd and 3rd sheets are Macro Enabled flash cards. The "New Question" button will randomize a question that corresponds with the provided times table. The answer will be populated, but the text format will keep it hidden. The "Show Answer" button will change the font color, and thus reveal the answer.